refactor(proteus): centralize `ProteusConversationSession` [WPB-15528]
